#2 abstracted out the home page to its own file out of main.dart
